docker/simple_api/requirements.txt
django-cors-headers==3.7.0
djangorestframework==3.12.4
